requirements-dev.txt
# --index-url https://download.pytorch.org/whl/cu124
# --extra-index-url https://pypi.org/simple
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
lightning==2.1.3
pyyaml
wandb
rdkit
ase
tqdm
cuequivariance_torch
mace-torch
scikit-learn
